// src/app/modules/core/models/nav-route.ts
export enum NavRoute {
  EMPTY_ROUTE = '/',
  FREE_TIPS = '/freeTips',
  TICKETS = '/tickets',
  PREMIUM_TIPS = '/premiumTips',
  ABOUT = '/about',
  ADD_TIP = '/admin/addTip',
  ADD_USER = '/admin/addUser',
  MANAGE_USERS = '/admin/manageUsers',
  SETTINGS = '/admin/settings',
  HISTORY = '/admin/history',
  PROFILE = '/profile',
  USER_LOGIN = '/userLogin',
  REGISTER = '/register',
  LOGOUT = '/userLogin',
  EDIT_ACTIVE_TIP = '/admin/editActiveTip',
  EDIT_NON_ACTIVE_TIP = '/admin/editNonActiveTip',
  SET_AS_FINISHED = '/admin/setAsFinished',
  EDIT_USER = '/admin/editUser',
  FORGOTTEN_PASSWORD = '/forgottenPassword',
  SANDBOX_GUIDE = '/guide'
}